eznarratives_reflex_kivy/requirements.txt
reflex>=0.3.6
kivy>=2.2.1
kivy-deps.sdl2>=0.6.0
kivy-deps.glew>=0.3.1
kivymd>=1.1.1
supabase>=2.0.0
openai>=1.3.0
python-dotenv>=1.0.0
psycopg2>=2.9.9
buildozer>=1.5.0
pyinstaller>=6.0.0
pillow>=10.0.0